Round and round.
include/innocent-pear/kthxbai.h :
- kthxbai_impl<...> can now try to scramble constants by
  rotating them a constant number of bits

include/innocent-pear/bbq.h :
- switched back from uint64_t to uint_least64_t
- added code to clip uint_least64_t values to 64 bits -- in
  case it can hold more
- pick_hi<.>(.) now only looks at lowest 64 bits of its
  argument
- moved ops_flags_t from innocent_pear::impl:: to
  innocent_pear::

include/innocent-pear/omg.h :
- now uses uint_least{64, 32, 16, 8}_t and uintptr_t instead
  of uint{64, 32, 16, 8}_t, even for machine specific code

include/innocent-pear/dawg.h ,
include/innocent-pear/yodawg.h ,
include/innocent-pear/host/lolcat.h ,
include/innocent-pear/host/moar.h ,
share/innocent-pear/lolrus.cc ,
share/innocent-pear/moar.cc :
- updated to match bbq.h

share/innocent-pear/doge-01.cc ,
share/innocent-pear/doge-02.cc ,
share/innocent-pear/doge-03.cc ,
share/innocent-pear/doge-04.cc ,
share/innocent-pear/doge-05.cc ,
share/innocent-pear/doge-06.cc ,
share/innocent-pear/doge-07.cc :
- updated to match bbq.h
- now uses `constexpr' instead of `static constexpr' for
  `flags' (and `flags2'), to prevent the flags from being
  emitted as data

bin/innocent-pear-dogecoin.cc ,
share/innocent-pear/omnomnom.lxx :
- uint64_t -> uint_least64_t; PRIx64 -> PRIxLEAST64

bin/innocent-pear-front-end.tcc ,
share/innocent-pear/calm.cc :
- removed unneeded explicit references to uint64_t

configure :
- instead of uint64_t, now tests for uint_least{64, 32, 16,
  8}_t, for host and target

GNUmakefile :
- updated to match ./configure

+++ include/innocent-pear/omg.h
@@ -103,17 +103,18 @@ struct omg_impl_0<State, 0u>
 #   if defined __arm__ && defined __thumb__
 			__asm __volatile(".inst.n %c0"
 			    : /* no outputs */
-			    : "n" ((uint32_t)pick_hi<uint16_t>(State^State2))
+			    : "n" ((uint_least32_t)pick_hi<uint_least16_t>
+				(State ^ State2))
 			    : "memory");
 #   elif defined __arm__
 			__asm __volatile(".inst %c0"
 			    : /* no outputs */
-			    : "n" (pick_hi<uint32_t>(State ^ State2))
+			    : "n" (pick_hi<uint_least32_t>(State ^ State2))
 			    : "memory");
 #   elif defined __i386__ || defined __amd64__
 			__asm __volatile(".byte %c0"
 			    : /* no outputs */
-			    : "n" (pick_hi<uint8_t>(State ^ State2))
+			    : "n" (pick_hi<uint_least8_t>(State ^ State2))
 			    : "memory");
 #   else
 			;
@@ -448,10 +449,10 @@ class omg
 				kthxbai<NewState, void *, Flags, Levels>
 				    p(&&foo, 1);
 #   ifdef __amd64__
-				uint64_t rzsz = 128 + 8 * Which7;
+				uintptr_t rzsz = 128 + 8 * Which7;
 #   endif
-				uint8_t x = static_cast<uint8_t>(State2 >> 24)
-				    / 2;
+				uint_least8_t x = static_cast<uint_least8_t>
+				    (State2 >> 24) / 2;
 				q = static_cast<void *>(p);
 				if (!q)
 					goto bar;
@@ -502,7 +503,7 @@ class omg
 					    innocent_pear_BR_PREFIX(5)
 					    "retq %3"
 					    : /* no outputs */
-					    : "r" (q), "g" ((uint64_t)x),
+					    : "r" (q), "g" ((uintptr_t)x),
 					      "r" (rzsz), "n" (rzsz),
 					      "n" (Which3), "n" (Which4)
 					    : /* no clobbers */
@@ -519,7 +520,7 @@ class omg
 					    "lretq %4"
 					    : /* no outputs */
 					    : "r" (q), "r" (r),
-					      "g" ((uint64_t)x), "r" (rzsz),
+					      "g" ((uintptr_t)x), "r" (rzsz),
 					      "n" (rzsz), "n" (Which3),
 					      "n" (Which4), "n" (Which5)
 					    : /* no clobbers */
@@ -554,7 +555,7 @@ class omg
 					    innocent_pear_BR_PREFIX(2)
 					    "retl"
 					    : /* no outputs */
-					    : "r" (q), "g" ((uint32_t)x),
+					    : "r" (q), "g" ((uint_least32_t)x),
 					      "n" (Which3)
 					    : /* no clobbers */
 					    : foo);  break;
@@ -568,8 +569,8 @@ class omg
 					    "lretl"
 					    : /* no outputs */
 					    : "r" (q), "r" (r),
-					      "g" ((uint32_t)x), "n" (Which3),
-					      "n" (Which4)
+					      "g" ((uint_least32_t)x),
+					      "n" (Which3), "n" (Which4)
 					    : /* no clobbers */
 					    : foo);  break;
 				    case 5:
@@ -703,7 +704,7 @@ class omg
 			if (!(Flags & allow_signal_safes)) {
 				union {
 					time_t t;
-					uint64_t i;
+					uint_least64_t i;
 				} u;
 				kthxbai<NewState, time_t *, Flags,
 				    Levels - 1> p(&u.t);
